fix: restructure the gpu file in prover
