+++ prover/src/gpu/mod.rs
@@ -1,0 +1,3 @@
+
+#[cfg(all(feature = "metal", target_arch = "aarch64", target_os = "macos"))]
+pub mod metal;
